requirements.txt
beautifulsoup4==4.10.0
black==21.12b0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.1
tomli==1.2.3
typing-extensions==4.0.1
urllib3==1.26.8
